+++ packages/miniprogram-element/CHANGELOG.md
@@ -1,5 +1,9 @@
 # 更新日志
 
+## 1.0.5
+
+* 调整 $$domInfo 对象的 class 为 className
+
 ## 1.0.4
 
 * 兼容 react 中 picker-view 设置数组到 value 中会被序列化字符串的问题